+++ lib/solidus_flexi_variants.rb
@@ -1,3 +1,5 @@
 require 'solidus_core'
+require 'solidus_support'
 require 'carrierwave'
 require 'solidus_flexi_variants/engine'
+require 'deface'
